Move GPU usage shenanigans into seperate file
