+++ src/utils/gpu/v3d.rs
@@ -1,5 +1,5 @@
 use anyhow::{Result, bail};
-use process_data::GpuIdentifier;
+use process_data::gpu_usage::GpuIdentifier;
 
 use std::path::{Path, PathBuf};
